+++ drizzle/0011_cloudy_kulan_gath.sql
@@ -1,0 +1,5 @@
+ALTER TABLE `demo` ADD `playback_status` text;--> statement-breakpoint
+ALTER TABLE `demo` ADD `playback_url` text;--> statement-breakpoint
+ALTER TABLE `demo` ADD `playback_pathname` text;--> statement-breakpoint
+ALTER TABLE `demo` ADD `playback_bytes` integer;--> statement-breakpoint
+ALTER TABLE `demo` ADD `playback_started_at` integer;